third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/block-container-block-end-nested-child-expected.txt

This is a testharness.js-based test.
[FAIL] item 2
  assert_equals: \n<item data-expected-margin-bottom="0" style="block-size: auto;">\n                <div><item data-expected-margin-bottom="10"></item></div>\n                <div>\n                    <item data-expected-margin-bottom="0" style="block-size: auto;">\n                        <div><item data-expected-margin-bottom="0"></item></div>\n                    </item>\n                    <item class="collapsed" data-expected-margin-bottom="0">\n                        <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>\n                    </item>\n                </div>\n                <item class="collapsed" data-expected-margin-bottom="0">\n                    <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>\n                </item>\n            </item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 4
  assert_equals: \n<item data-expected-margin-bottom="0" style="block-size: auto;">\n                        <div><item data-expected-margin-bottom="0"></item></div>\n                    </item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 5
  assert_equals: \n<item data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 6
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0">\n                        <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>\n                    </item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 7
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 8
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0">\n                    <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>\n                </item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 9
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 10
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0">\n                <div><item class="collapsed" data-expected-margin-bottom="0"></item></div>\n            </item>\nmargin-bottom expected "0" but got "10"
[FAIL] item 11
  assert_equals: \n<item class="collapsed" data-expected-margin-bottom="0"></item>\nmargin-bottom expected "0" but got "10"
Harness: the test ran to completion.
